fix(scene3d): unbreak the Linux build and clear the module's warnings
meshShadowBounds is [[nodiscard]], and the new seam test dropped its result.
GCC with -Werror rejects that outright, so the Linux build failed on a test
that passes here — the macOS build only WARNED, because the bring-up-era
if(APPLE) branch strips -Werror, and a warning is easy to miss in a build log.
The return value is now asserted rather than voided: the caster bounds come
from the same list the test just checked was pushed, so a valid box is what
makes that push meaningful rather than incidental.

The remaining eight are the same class, found by replaying the module's
compiles with -Werror restored. All are value-preserving:

Six are half-done casts — an expression where someone cast one operand and
left the next to convert implicitly (static_cast<size_t>(y) * w,
static_cast<std::size_t>(divisions) * divisions). The cast had already been
judged necessary; it was just applied incompletely. arrow_mesh.cpp is the
clearest case: every LATER index loop in that same file already casts
explicitly, so those four lines were stragglers rather than a house style.

occupancy_grid_reconstructor_test is subtler: kGrid is uint32_t, so
(i * 3) % (kGrid - 2) was converting i * 3 to unsigned. Keeping the whole
expression signed yields identical values for positive i.

pointcloud_layer is the only one that is not a cast. Each range-spinbox
handler captured BOTH spin boxes but only ever touches its sibling — the
range_min handler clamps range_max and vice versa — so each was capturing
itself for nothing.

+++ pj_scene3D/tests/occupancy_grid_reconstructor_test.cpp
@@ -281,8 +281,8 @@ TEST(OccupancyGridReconstructorTest, DuplicateTimestampSnapshotBoundary) {
   // 64th update — where the stride-driven snapshot fires — falls mid-group.
   for (int i = 1; i <= 100; ++i) {
     const Timestamp ts = 1000 + 10 * ((i - 1) / 5 + 1);
-    const int32_t x = (i * 3) % (kGrid - 2);
-    const int32_t y = (i * 7) % (kGrid - 2);
+    const int32_t x = (i * 3) % (static_cast<int32_t>(kGrid) - 2);
+    const int32_t y = (i * 7) % (static_cast<int32_t>(kGrid) - 2);
     tl.updates.push_back(makeUpdate(ts, x, y, 2, 2, std::vector<uint8_t>(4, static_cast<uint8_t>(i))));
   }
   const auto bp = baseProviderFor(tl.bases);
